break out the big cypher extension into logically grouped partials
